refact: rename expense/model/types to expense/model
