+++ src/app/routes/expenses.$uid.index.tsx
@@ -7,7 +7,7 @@ import {
   useExpenseByUid,
   useUpdateExpense,
 } from '@/features/expense/api/useExpenseQuery';
-import { Expense } from '@/features/expense/model/types/Expense';
+import { Expense } from '@/features/expense/model/Expense';
 import ExpenseForm from '@/features/expense/ui/ExpenseForm';
 import SubPageHeader from '@/shared/ui/SubPageHeader';
