+++ src/features/expense/ui/DailyExpenseRecord/DailyExpenseRecord.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import { describe, it } from 'vitest';
 
-import { Expense } from '@/features/expense/model/types/Expense';
+import { Expense } from '@/features/expense/model/Expense';
 
 import DailyExpenseRecord from './DailyExpenseRecord';
