+++ src/features/expense/ui/ReviewExpenseCard/ReviewExpenseCard.tsx
@@ -2,7 +2,7 @@ import { format } from 'date-fns';
 import { useLayoutEffect, useRef, useState } from 'react';
 
 import CategoryTag from '@/features/category/ui/CategoryTag';
-import { Expense } from '@/features/expense/model/types/Expense';
+import { Expense } from '@/features/expense/model/Expense';
 import ArrowLeft from '@/shared/ui/icons/ArrowLeft';
 import { cn } from '@/shared/ui/styles/utils';
