+++ src/features/expense/ui/SelectMonthList/SelectMonthList.tsx
@@ -1,4 +1,4 @@
-import Period from '@/features/expense/model/types/Period';
+import Period from '@/features/expense/model/Period';
 
 import SelectMonthListItem from '../SelectMonthListItem';
