+++ src/features/expense/utils.ts
@@ -1,5 +1,6 @@
+import { ConsumptionKind } from '@/features/expense/model/ConsumptionKind';
+
 import { consumptionMap, consumptionTextsMap } from './consts';
-import { ConsumptionKind } from './model/types/ConsumptionKind';
 
 export function getConsumptionTitle(kind: ConsumptionKind): string {
   // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
